Rename vars to be more generic for table creation
