+++ pkg/jobrunaggregator/cmd.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/openshift/ci-tools/pkg/jobrunaggregator/jobrunaggregatoranalyzer"
 	"github.com/openshift/ci-tools/pkg/jobrunaggregator/jobrunbigqueryloader"
-	"github.com/openshift/ci-tools/pkg/jobrunaggregator/jobruntablecreator"
 	"github.com/openshift/ci-tools/pkg/jobrunaggregator/jobtableprimer"
 	"github.com/openshift/ci-tools/pkg/jobrunaggregator/releasebigqueryloader"
+	"github.com/openshift/ci-tools/pkg/jobrunaggregator/tablescreator"
 )
 
 func NewJobAggregatorCommand() *cobra.Command {
@@ -25,6 +25,6 @@ func NewJobAggregatorCommand() *cobra.Command {
 	cmd.AddCommand(releasebigqueryloader.NewBigQueryReleaseTableCreateFlagsCommand())
 	cmd.AddCommand(releasebigqueryloader.NewBigQueryReleaseUploadFlagsCommand())
 
-	cmd.AddCommand(jobruntablecreator.NewBigQueryJobRunTableCreateFlagsCommand())
+	cmd.AddCommand(tablescreator.NewBigQueryCreateTablesFlagsCommand())
 	return cmd
 }
